refactor(clawdstrike-policy-event): extract honey from edr/mod.rs
Move the honey-artifact cluster (observed_path, honey_artifact_match_
evidence, observation_contains_marker, serialized_contains_marker,
honey_artifact, honey_contents, ensure_safe_relative_path,
create_new_honey_file, and the cfg'd set_file_permissions pair) into
edr/honey.rs. The cfg(unix)/cfg(not(unix)) set_file_permissions
attributes move verbatim.

deception.rs and detection/supply_chain.rs reach honey_artifact,
honey_artifact_match_evidence, ensure_safe_relative_path, and
create_new_honey_file via super:: paths, so those four are re-exported
pub(crate) at the edr root. Pure relocation; bodies unchanged.
